social/requirements.txt
asgiref==3.7.2
Django==3.2.19
django-filter==23.2
djangorestframework==3.14.0
drf-nested-routers==0.93.4
importlib-metadata==6.6.0
Markdown==3.4.3
Pillow==9.5.0
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.2
zipp==3.15.0
